chore: remove DEVELOPER_MODE and apply clang-tidy fixes
- Remove FTXUI_DEVELOPER_MODE option and its logic from CMakeLists.txt.
- Disable readability-math-missing-parentheses in .clang-tidy.
- Apply various clang-tidy improvements (emplace_back, const refs, at(), etc.).

+++ src/ftxui/dom/box_helper.cpp
@@ -12,8 +12,9 @@ namespace ftxui::box_helper {
 namespace {
 
 int SafeRatio(int value, int numerator, int denominator) {
-  return static_cast<int64_t>(value) * static_cast<int64_t>(numerator) /
-         std::max(static_cast<int64_t>(denominator), static_cast<int64_t>(1));
+  return static_cast<int>(
+      static_cast<int64_t>(value) * static_cast<int64_t>(numerator) /
+      std::max(static_cast<int64_t>(denominator), static_cast<int64_t>(1)));
 }
 
 // Called when the size allowed is greater than the requested size. This
